Folder structure: 1.Created a modular structure to separate concerns at path src/modules
2. Shared items like layouts moved to /src/shared folder
