+++ src/modules/Charts/views/ChartReview.jsx
@@ -1,0 +1,5 @@
+export const ChartReview = () => {
+    return (
+        <></>
+    )
+}
