+++ src/routes.js
@@ -1,6 +1,10 @@
+import {ChartReview} from "./modules/Charts/views/ChartReview";
 
 const routes = [
-
+    {
+        path: "/charts/review",
+        component: ChartReview
+    }
 ];
 
 export default routes;
