+++ src/shared/pages/Home.js
